Improve downloading algorithm

+++ scripts/R/run_notebook.sh
@@ -42,28 +42,40 @@ URL2="https://vastdb.crg.eu/downloads/mm10/EVENT_INFO-mm10.tab.gz"
 FILE1="$PWD/notebooks/final/SPLICE_SITE_SCORES-mm10.tab.gz"
 FILE2="$PWD/notebooks/final/EVENT_INFO-mm10.tab.gz"
 
+UNZIPPED_FILE1="${FILE1%.gz}"
+UNZIPPED_FILE2="${FILE2%.gz}"
+
 # Create the directory if it doesn't exist
 mkdir -p "$PWD/notebooks/final"
 
-# Check and download the first file
-if [ ! -f "$FILE1" ]; then
-    echo "$FILE1 not found. Downloading..."
-    wget "$URL1" -O "$FILE1"
+# Check and download the first file if the unzipped version doesn't exist
+if [ ! -f "$UNZIPPED_FILE1" ]; then
+    if [ ! -f "$FILE1" ]; then
+        echo "$FILE1 not found. Downloading..."
+        wget "$URL1" -O "$FILE1"
+    else
+        echo "$FILE1 already exists. Skipping download."
+    fi
+    echo "Unzipping $FILE1..."
+    gunzip -c "$FILE1" > "$UNZIPPED_FILE1"
 else
-    echo "$FILE1 already exists. Skipping download."
+    echo "$UNZIPPED_FILE1 already exists. Skipping download and unzip."
 fi
 
-# Check and download the second file
-if [ ! -f "$FILE2" ]; then
-    echo "$FILE2 not found. Downloading..."
-    wget "$URL2" -O "$FILE2"
+# Check and download the second file if the unzipped version doesn't exist
+if [ ! -f "$UNZIPPED_FILE2" ]; then
+    if [ ! -f "$FILE2" ]; then
+        echo "$FILE2 not found. Downloading..."
+        wget "$URL2" -O "$FILE2"
+    else
+        echo "$FILE2 already exists. Skipping download."
+    fi
+    echo "Unzipping $FILE2..."
+    gunzip -c "$FILE2" > "$UNZIPPED_FILE2"
 else
-    echo "$FILE2 already exists. Skipping download."
+    echo "$UNZIPPED_FILE2 already exists. Skipping download and unzip."
 fi
-# Unzip the downloaded files
-echo "Unzipping files..."
-gunzip -c "$FILE1" > "${FILE1%.gz}"
-gunzip -c "$FILE2" > "${FILE2%.gz}"
+
 # Set the working directory inside the container to /workspace
 singularity run --bind "$(pwd)/notebooks/final:/shared" \
   docker://andresgordoortiz/splicing_analysis_r_crg:v1.3 \
